requirements.txt
requests
ollama
streamlit
pandas
openai
google-generativeai